+++ src/main/java/devkor/com/teamcback/domain/admin/route/validation/CommaNumericValidator.java
@@ -1,4 +1,4 @@
-package devkor.com.teamcback.domain.admin.validation;
+package devkor.com.teamcback.domain.admin.route.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
